Tweaked money donation page. TODO: Page title, Appbar Logo, and separate titleWidget
